+++ src/plugins/economy/merch.py
@@ -259,12 +259,12 @@ async def post_merch_transfer(request: Request, response: Response, itemid: int,
         response.status_code = 400
         return {"error": ml.tr(request, "invalid_owner", force_lang = au["language"])}
 
-    await app.db.execute(dhrid, f"SELECT userid FROM economy_merch WHERE itemid = {itemid}")
+    await app.db.execute(dhrid, f"SELECT merchid, userid FROM economy_merch WHERE itemid = {itemid}")
     t = await app.db.fetchall(dhrid)
     if len(t) == 0:
         response.status_code = 428
         return {"error": ml.tr(request, "merch_not_found", force_lang = au["language"])}
-    current_owner = t[0][0]
+    (merchid, current_owner) = (t[0][0], t[0][1])
     if current_owner == foruser:
         response.status_code = 409
         return {"error": ml.tr(request, "new_owner_conflict", force_lang = au["language"])}
@@ -282,6 +282,24 @@ async def post_merch_transfer(request: Request, response: Response, itemid: int,
     await app.db.execute(dhrid, f"UPDATE economy_merch SET userid = {foruser} WHERE itemid = {itemid}")
     await app.db.execute(dhrid, f"INSERT INTO economy_transaction(from_userid, to_userid, amount, note, message, from_new_balance, to_new_balance, timestamp) VALUES ({current_owner}, {foruser}, NULL, 'm{itemid}-transfer', '{convertQuotation(message)}', NULL, NULL, {int(time.time())})")
     await app.db.commit(dhrid)
+
+    from_user = await GetUserInfo(request, userid = current_owner)
+    to_user = await GetUserInfo(request, userid = foruser)
+    from_user_language = await GetUserLanguage(request, from_user["uid"])
+    to_user_language = await GetUserLanguage(request, to_user["uid"])
+
+    from_message = ""
+    to_message = ""
+    if message != "":
+        from_message = "  \n" + ml.tr(request, "economy_transaction_message", var = {"message": message}, force_lang = from_user_language)
+        to_message = "  \n" + ml.tr(request, "economy_transaction_message", var = {"message": message}, force_lang = to_user_language)
+
+    merch = ml.ctr(request, "unknown") + " (" + merchid + ")"
+    if merchid in app.merch.keys():
+        merch = app.merch[merchid]["name"]
+
+    await notification(request, "economy", from_user["uid"], ml.tr(request, "economy_sent_transaction_item", var = {"type": "1x " + ml.tr(request, "merch", force_lang = from_user_language).title(), "name": merch, "to_user": to_user["name"], "to_userid": to_user["userid"] if to_user["userid"] is not None else "N/A", "message": from_message}, force_lang = from_user_language))
+    await notification(request, "economy", to_user["uid"], ml.tr(request, "economy_received_transaction_item", var = {"type": "1x " + ml.tr(request, "merch", force_lang = from_user_language).title(), "name": merch, "from_user": from_user["name"], "from_userid": from_user["userid"] if from_user["userid"] is not None else "N/A", "message": to_message}, force_lang = to_user_language))
 
     return Response(status_code=204)